Debugging length modifier specs

+++ print_int.c
@@ -10,11 +10,11 @@
 int print_int(va_list integer, flag *f)
 {
 	int length = 0;
-	int num;
+	long int num;
 	unsigned long int n = 0;
 
-	n = va_arg(integer, unsigned long int);
-	num = (long int)n;
+	n = va_arg(integer, unsigned int);
+	num = n;
 
 	if (f->plus && num >= 0)
 		length += _putchar('+');
